+++ src/day8/re_frame/trace/subs.cljs
@@ -225,6 +225,21 @@
       match)))
 
 (rf/reg-sub
+  :epochs/current-match-state
+  :<- [:epochs/epoch-root]
+  :<- [:epochs/match-ids]
+  (fn [[epochs match-ids] _]
+    (let [current-id (:current-epoch-id epochs)
+          match      (cond
+                       (nil? current-id) (last (:matches-and-subs epochs))
+                       (< current-id (first match-ids)) (first (:matches-and-subs epochs))
+                       ;; This case seems impossible, but can happen if the user filters out
+                       ;; an event that they are 'on'.
+                       (> current-id (last match-ids)) (last (:matches-and-subs epochs))
+                       :else (get (:matches-by-id2 epochs) current-id))]
+      match)))
+
+(rf/reg-sub
   :epochs/current-event-trace
   :<- [:epochs/current-match]
   (fn [match _]
@@ -322,9 +337,14 @@
           [start end] (first (drop (dec frame-number) frames))]
       (metam/elapsed-time start end))))
 
-
-
 (rf/reg-sub
+  :timing/event-processing-time
+  :<- [:epochs/current-match-state]
+  (fn [match]
+    (utils/spy "match" match)
+    (get-in match [:timing :re-frame/event-time])))
+
+#_(rf/reg-sub
   :timing/event-processing-time
   :<- [:traces/current-event-traces]
   (fn [traces]