+++ CHANGELOG.md
@@ -26,6 +26,7 @@ This version requires re-frame 0.10.4 to make use of the newly added Event panel
 * Layer 2/3 calculations are more accurate now. We now use the last seen layer level when a subscription runs, to inform it's layer level if it was created or destroyed.
 * View namespaces that are ignored are no longer shown when showing traces for all epochs.
 * Distinguish between subscriptions that return `nil` values and those that haven't run yet.
+* Timing panel not showing elapsed event processing time.